Fix v8 deopt problem
`hash` can be either `Hash` object or `undefined` in those two lines.
V8 will create a binary for `hash`, which is `Hash` object, to optimize
the code. But when it found that `hash` is `undefined`, it'll have to
remove that binary and create a new one (deopt). This process cause the
app to lose some performance.

With this change, `hash` won't be used when it's `undefined` so the
compiler won't have to deopt.